requirements.txt
requests
pandas
beautifulsoup4
boto3
python-telegram-bot
datetime
